pipelines/rj_iplanrio__cor_alerts_test_writer/pyproject.toml
[project]
name = "rj_iplanrio__cor_alerts_test_writer"
version = "0.1.0"
description = "Pipeline para geracao de dados de teste do COR alerts aggregator"
dependencies = [
    "httpx>=0.27.0",
    "pandas>=2.0.0",
    "google-cloud-bigquery>=3.25.0",
    "basedosdados==2.0.0b23",
]
